pyproject.toml
[tool.poetry]
name = "wiki-watch"
version = "0.1.0"
description = ""
authors = ["Charlie Cumber <chukc@hotmail.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.1"
flask-sqlalchemy = "^3.0.3"
flask-migrate = "^4.0.4"
python-dotenv = "^1.0.0"
mysql = "^0.0.3"
websockets = "^11.0.2"
pytz = "^2023.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
